Make remove_virtual_function() consistent
If there is more than one possible virtual function, we cast the object
to the appropriate type in each case, but if there is only one
possibility we don't. This PR makes us cast the object when there is
only one possibility as well.

Note that just because there is only one possibility, it does not mean
that the object already has the correct type. Also, if the object is
already the correct type then the cast isn't inserted, in all cases.