[TG-2585] Support implicitly generic superclasses
Before only generic superclasses where supported for the java generic symbols,
as the detection was to match on the suffix `>;`. This commit changes it to
match an opening `<` in the superclass reference in order to detect a generic
superclass which will also succeed in the case of an implicitly generic one.