Add callback after symex to bmct
This permits frontend programs to intervene after symex executes, either in addition to
or instead of running bmc. Its current use case is to implement show-goto-functions and
similar debug output options when symex-driven lazy loading means that function bodies are
not populated until symex has run.

When --paths is in use, the callback will be made every time symex finishes a path.