JBMC: use symex-driven lazy loading
This adds the symex-driven-lazy-loading option, which causes functions to be loaded only when
symex first requests their body. This means that all frontend effort can be saved relating to
functions that symex reveals to be inaccessible, such as virtual function calls whose full array
of possible targets are known to be unreachable as a result of value-set analysis on the type of
the callee instance.