Avoid double invoking ccache
In Travis, when jobs have the 'cache: ccache' option set, Travis
automatically adds the location of the ccache wrappers around g++/clang++
to the head of the PATH environment variable. This means that when you
invoke 'g++' or 'clang++' what you are actually invoking is ccache.
However, our jobs execute 'ccache g++', which means we are actually
invoking ccache on itself. This was causing very poor ccache hit rates.