Increase ccache size for debug builds
Currently we use a 1GB limit on the ccache size for all builds.
However for the debug builds (those specifying '-g' to g++) object
sizes are significantly larger due to the extra DWARF debug info.
This is causing the debug builds to hit the ccache size limit.
Local experiments suggest that with a 1GB limit, we can only get
about 80% hit rate. Increasing the ccache size to 2GB increases
the hit rate to 99%.