Continue parsing after unsupported  BootstrapMethods entry
Before it was possible to keep some `BootstrapMethods` entries unparsed, after a
currently unsupported was found. This lead to the situation that JBMC ended up
with a wrong file offset which caused problems in the case of class file
attributes being present after the `BootstrapMethods` attribute.

This was not observed using the Oracle or OpenJDK compilers, but currently only
on the Eclipse compiler for Java.