Moved the first condition to after the explanation of our assumptions
Here it is clearer why we need at least 3 arguments. Also documented
this explictly above the condition. Also changed to a debug as there is
nothing in the java spec requiring entries to have at least 3
parameters, it is simpyl our assumption about what entries can appear in
the bootstrap methods table