goto-analyzer (un)reachable-functions: build valid json output
Some functions do not have a location attached to the END_FUNCTION instruction.
This should be fixed (as is discussed in PR#918), but until such a fix is in
place this workaround will ensure JSON output is valid by not exclusively
relying on the END_FUNCTION location.

The regression tests now explicitly test for absence of such invalid JSON.