Fix broken static_init_order test descriptions
These were trying to match for a blank line, which doesn't occur in the output
any more, and was surely not intentionally searched for in the first place. This
only causes a problem under Windows (or more likely, under certain versions of Perl),
probably due to a change in empty line stripping.