Use a const expr to avoid unnecessary detach
While dest.type() is passed as const reference, dest itself is not const
in this context and thus irept::add will be used instead of irept::find.
This, in turn, results in unnecessary calls to detach and thus breaks
sharing and causes memory allocation.

This change reduces linking time in some Linux kernel experiments by a
further 10%.