Add ability to overlay classes with new definitions of existing methods
Allows multiple definitions of the same class to appear on the classpath, so long as all but the first definition are marked with the attribute @java::com.diffblue.security.OverlayClassImplementation. Overlay class definitions can contain methods with the same signature as methods in the original class, so long as these are marked with the attribute @java::com.diffblue.security.OverlayMethodImplementation; such overlay methods are replaced in the original file with the version found in the last overlay on the classpath. Later definitions can also contain new supporting methods and fields that are merged in.
This will allow insertion of Java methods into library classes to handle, for example, modelling dependency injection.
Includes regression tests
Improvements over earlier version:
  Allows loading of classes such as java.lang.Object from an explicit model
  Prevent attempting to load any class more than once
  Invalid values in the classpath are ignored
  Handles duplicate files appearing in multiple items on the classpath
  Filter loaded class files to remove duplicates or misordered overlays before following links to other classes so as not to add dependencies of duplicate classes