Can now test for an option being set in optionst
There was previously no public interface to check whether an option had
been set in an optionst object; this meant that clients were instead
using the isset() method of cmdlinet. This change allows us to set the
options from the cmdlinet near the beginning of the front-end, and not
refer to the cmdlinet afterward.