Attribute main function arguments to __CPROVER_start
These are created to serve as arguments to main() (or whatever --function was passed), but are
declared in __CPROVER_start, and so should be named after their declarer like all other locals.

This was causing a problem in the security repository, where we relied on local variable names
being prefixed by their declaring function name, and the easiest fix was to restore that convention.