Make get_subexpression_at_offset return an exprt for consistency
All other functions declared in that block in pointer_offset_size.h return an
exprt and the comment explicitly states that they return nil_exprt() on failure.
get_subexpression_at_offset used to return "true" on success (and false on
failure), which is the opposite of what most of the code base does. (But the
interpretation of true/false as return values is non-obvious anyway and thus
returning an expression adds clarity.)