Generalise option setting from strategy unit tests
The check_with_strategy methods previously took an argument to set the
`unwind-limit` option with. This commit replaces that argument with a
function that mutates the optionst that is passed into it. This can be
used for setting `unwind-limit` or any other option from the client.
This is in preparation for a new strategy that requires additional
options to be set.