called_functions: use unordered_set
Instead of using an ordered set, use a faster unordered set. This way,
running the algorithms against larger code bases should result in
decreased run times, while on small programs the run time might slightly
increase.

As this function is used in function pointer removal, and in code that
computes unreachable functions, the type change had to be propagated to
the related classes as well.