Object factory: permit null pointers within java.lang.Class
This is in fact required by the Deeptest models library, and is generally sensible if we don't
know whether or not a particular implementation of java.lang.Class should or should not be able
to store null in any particular field. If you're running an implementation that requires them
not null, add a cproverNondetInitialize function to that effect.

This was blocking using implementations that expect a null value from passing initialization.