function-pointer-removal: drop unused set
The compute_called_functions method maintains two sets, 'done' and
'functions'. The set 'done' is only used to determine whether an element
can be skipped. Since the content of the two sets is the same, the set
'done' can be dropped, and 'functions' can be used instead. Hence, this
commit drops the set 'done'.

Furthermore, the return value of insert includes a bool that indicates
the success of the insert operation. In case insertion failed, the element
has been present already, so that we can merge the find+insert into a
single insert operation.

Fixes: #1783