get_max: do not parse every symbol name in the symbol table
The typical case is that we have a small number of entries with the same
prefix. Just perform that small number of name lookups (which are
constant time after the log-time conversion from a string to an
irep_idt). This speeds up function pointer removal on a build of the
Linux kernel from 4610 seconds to just 86 seconds.

Also rename get_max to smallest_unused_suffix to disambiguate the case
of a symbol not being present, which in turn simplifies the use.