Add support for Java annotations with Class value
Java annotations can store element-value pairs, where each value is of
type String, Enum, Class, annotation, array or a primitive type.
The Class case was among the cases that did not have support in the java
bytecode parser before. For minimal support (without storing any
reference to java.lang.Class), we can just store the name of the type as
retrieved from the constant pool in a symbol_exprt.