Restore _Float128 support by default
_Float128 was exposed by GCC as of version 4.3 (x86-64) to 4.5 (i386 and I think others).
cbmc still makes no effort to imitate a particular GCC version, so I simply enable recognising
the token by default, which used to happen universally; goto-cc on the other hand uses the
new `gcc_versiont` to choose an appropriate behaviour.

This fixes a regression on tests using math.h due to some versions of glibc defining _Float128
conditional on the compiler version, which cbmc and goto-cc now both allow to appear like the
gcc found on the path.