Reduce memory cost of test by using a scalar variable instead of an array
The test previously required 2.6GB of memory to complete, and was sometimes
getting killed (seemingly only in the CMake configuration, though). Just adding
them up in a scalar variable seems to suffice to prove the effectiveness of the
original bugfix in 1b244db4ec13d868d7b243. It did, however, take adding one more
variable to actually cause an invariant failure - with 256 variables the sum was
just wrong.