Documentation wording for replace function bodies
Previous use of the word "implementation" is misleading, as "implementation" is synonymous with "function body".  Here "implementation" is intended to refer to the derived class used.

The previous wording could be interpreted as "replace the function body with one based on the function body, i.e., a summary of the function body code. This will work the same regardless of whether or not the function already has a body".

Documentation rewording: generate_function_body

This function cannot behave the same whether or not the function passed to it has a function body. The implementation of "generate_function_body" contains a PRECONDITION that the function body is not available.