Make it clearer that <=> will always be stringified
clang++-6.0 warns that in the next C++ standard <=> will be an operator and therefore an atomic
lexical token. I think that's overly zealous, but it's easily worked around by explicitly quoting
the name, which will be stringified before use in any case.

See https://github.com/diffblue/cbmc/issues/1974 for more details, or
https://bugs.llvm.org/show_bug.cgi?id=36925 for an upstream bug report.