Replace lambda by member function reference to silence Visual Studio
Visual Studio previously warned about a reference to a local or temporary being
returned. This seems spurious, possibly caused by "this" being passed by-value
to the lambda and Visual Studio thus assuming a copy was being created. Going
via a reference to a member function makes Visual Studio happy (and should not
change the behaviour in any way).