Add useful functions to class_hierarchy_grapht
class_hierarchy_grapht is a more advanced representation of the class
hierarchy than class_hierarchyt. So ideally we want to migrate all
functions from the old version to the graph-based representation. This
commit does this for the functions get_parents_trans,
get_children_trans, and introduces a new function get_direct_children.