Remove INVARIANT in full_struct_abstract_object copy constructor
There are times during the process of writing to a struct that
the copy constructor gets called (via mutable_clone) while the
source object is in an 'in-flight' state which gets discarded by
the time the full write actions have been completed. Leaving the
INVARIANT check in-place causes some writes to fail because of the
transiently inconsistent state of the full_struct_abstract_object.